modificacion de database add .env
