+++ src/config/database.js
@@ -1,0 +1,21 @@
+const mysql = require('mysql2');
+require('dotenv').config();
+
+// Crear la conexión a la base de datos
+const connection = mysql.createConnection({
+  host: process.env.DB_HOST,
+  user: process.env.DB_USER,
+  password: process.env.DB_PASSWORD,
+  database: process.env.DB_NAME
+});
+
+// Conectar a la base de datos
+connection.connect((err) => {
+  if (err) {
+    console.error('Error conectándose a la base de datos:', err.stack);
+    return;
+  }
+  console.log('Conectado a la base de datos como id ' + connection.threadId);
+});
+
+module.exports = connection;
